backend finished and front end auth

+++ Backend/src/controllers/notes.ts
@@ -2,11 +2,14 @@ import { RequestHandler } from "express"
 import NoteModel from "../models/note"
 import createHttpError from "http-errors"
 import mongoose from "mongoose"
+import { assertIsDefined } from "../util/assertIsDefined"
 
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId
     try {
-        const notes = await NoteModel.find().exec()
+        assertIsDefined(authenticatedUserId)
+        const notes = await NoteModel.find({userId:authenticatedUserId}).exec()
         res.status(200).json(notes)
     } catch (error) {
         next(error)
@@ -15,14 +18,19 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async(req,res,next) =>{
     const id = req.params.noteId
-    
+    const authenticatedUserId = req.session.userId
     try {
+        assertIsDefined(authenticatedUserId)
         if(!mongoose.isValidObjectId(id)){
             throw createHttpError(400,"el ID es invalido")
         }
         const note = await NoteModel.findById(id).exec();
         if(!note){
             throw createHttpError(404,"no se ha encontrado una nota con ese ID")
+        }
+
+        if(note.userId === authenticatedUserId){
+            throw createHttpError(401,"you cannot access this note")
         }
         res.status(200).json(note)
     } catch (error) {
@@ -39,11 +47,14 @@ export const createNote: RequestHandler<unknown,unknown,CreateNoteBody,unknown> 
 
     const title = req.body.title
     const description = req.body.description
+    const authenticatedUserId = req.session.userId
     try {
+        assertIsDefined(authenticatedUserId)
         if(!title || title.trim() === ""){
             throw createHttpError(400,"No se ha proveido un Titulo")
         }
         const newNote = await NoteModel.create({
+            userId:authenticatedUserId,
             title,
             description,
         })
@@ -66,7 +77,9 @@ export const updateNote : RequestHandler<UpdateNoteParams,unknown,UpdateNoteBody
     const id = req.params.noteId
     const title = req.body.title
     const description = req.body.description
+    const authenticatedUserId = req.session.userId
     try {
+        assertIsDefined(authenticatedUserId)
         if(!mongoose.isValidObjectId(id)){
             throw createHttpError(400,"el ID es invalido")
         }
@@ -78,6 +91,10 @@ export const updateNote : RequestHandler<UpdateNoteParams,unknown,UpdateNoteBody
         const note = await NoteModel.findById(id).exec();
         if(!note){
             throw createHttpError(404,"no se ha encontrado una nota con ese ID")
+        }
+
+        if(note.userId === authenticatedUserId){
+            throw createHttpError(401,"you cannot access this note")
         }
         
         note.title = title
@@ -91,13 +108,19 @@ export const updateNote : RequestHandler<UpdateNoteParams,unknown,UpdateNoteBody
 
 export const deleteNote: RequestHandler = async (req,res,next) => {
     const id = req.params.noteId;
+    const authenticatedUserId = req.session.userId
     try {
+        assertIsDefined(authenticatedUserId)
         if(!mongoose.isValidObjectId(id)){
             throw createHttpError(400,"el ID es invalido")
         }
         const note = await NoteModel.findById(id)
         if(!note){
             throw createHttpError(404,"Nota no Encontrada")
+        }
+
+        if(note.userId === authenticatedUserId){
+            throw createHttpError(401,"you cannot access this note")
         }
 
         await note.deleteOne()
